Move camera postion to always see the project in the window
I attribuated a box of focus and make sure this box is alway fully visible on resize.
To do so I update the depth position of the camera to always match the focusBox.
I also move  the project in the center of the screen if it's a mobile device.
And if it's on desktop I move the X Axis of the camera
so the focusBoxLeft egde is matching the screen left edge.
Export 'ControlsDebug' class as default to see what I mean.