Add responsive tablet+mobile
I change the size of the project focus box to not be crop on small device.
I added  breakpoints in there own file.
I move the stats values position from the side to bellow,
on tablet and lower.
On Tablet and lower the project first impression text
is now filling the all width of the window