Now Mongoose services return POJOs
In fact mongoose queries are not return javascript object,
but kind of like a Mongoose document I think,
anyway the point is that you have more than your requested for,
so to avoid this I have to convert my query returns to POJOs.
A POJO is a plain javascript object that only contain data and no method.
I added a schema interface, containing only the properties that I want in my object.
I also added a Document interface for each model that represent the result of the mongoose queries.

ALL OF THIS IS FOR HAVING A STRONGLY TYPE API