Add ProjectApi proxy to Portfolio app
This allow the portfolio to access the Project API by using OwnUrl + '/api/projects'.
Note that /api/projects is removed from the request.
If you want to access the route /HugoMeet of the ProjectApi you should send your resquest to OwnUrl + '/api/projects/HugoMeet'.
'/api/projects' is just a prefix to know to which API send the request to.