Convert existing route to return Stats has IStat/IStats
instead of returning the full models.
This also mean that when returning multiple stats you are not returning an array anymore but an hashtable so it's easyer to use this data on the client side.

In the mean time I update my Portfolio component who were using this routes to show the data has expected.